upbit_alert_bot/requirements.txt
python-telegram-bot==13.7
pyJWT==2.1.0
requests==2.26.0